+++ src/EditorFeatures/CSharpTest/Classification/TotalClassifierTests.cs
@@ -6,11 +6,17 @@ using System.Collections.Immutable;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Classification;
+using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Classification;
+using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
+using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Remote.Testing;
 using Microsoft.CodeAnalysis.Shared.Extensions;
+using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.CodeAnalysis.Text;
+using Microsoft.CodeAnalysis.Text.Shared.Extensions;
+using Microsoft.VisualStudio.Text;
 using Roslyn.Test.Utilities;
 using Roslyn.Utilities;
 using Xunit;
@@ -3021,6 +3027,119 @@ Keyword("async"));
                 Punctuation.CloseAngle,
                 Local("fp"),
                 Punctuation.Semicolon);
+        }
+
+        [WpfFact]
+        public async Task TestTotalClassifier()
+        {
+            using var workspace = TestWorkspace.CreateCSharp(""""
+                using System.Text.RegularExpressions;
+
+                class C
+                {
+                    // class D { }
+                    void M()
+                    {
+                        new Regex("(a)");
+                        var s1 = "s1";
+                        var s2 = $"s2";
+                        var s3 = @"s3";
+                        var s4 = """
+                        s4
+                        """;
+                    }
+                }
+                """");
+            var document = workspace.Documents.First();
+
+            var listenerProvider = workspace.ExportProvider.GetExportedValue<IAsynchronousOperationListenerProvider>();
+            var globalOptions = workspace.ExportProvider.GetExportedValue<IGlobalOptionService>();
+
+            var provider = new TotalClassificationTaggerProvider(
+                workspace.GetService<IThreadingContext>(),
+                workspace.GetService<ClassificationTypeMap>(),
+                globalOptions,
+                visibilityTracker: null,
+                listenerProvider);
+
+            var buffer = document.GetTextBuffer();
+            using var tagger = provider.CreateTagger(document.GetTextView(), buffer);
+
+            var waiter = listenerProvider.GetWaiter(FeatureAttribute.Classification);
+            await waiter.ExpeditedWaitAsync();
+
+            var allCode = buffer.CurrentSnapshot.GetText();
+            var tags = tagger!.GetTags(new NormalizedSnapshotSpanCollection(buffer.CurrentSnapshot.GetFullSpan()));
+
+            var actualOrdered = tags.OrderBy((t1, t2) => t1.Span.Span.Start - t2.Span.Span.Start);
+
+            var actualFormatted = actualOrdered.Select(a => new FormattedClassification(allCode.Substring(a.Span.Span.Start, a.Span.Span.Length), a.Tag.ClassificationType.Classification));
+
+            AssertEx.Equal(new[]
+            {
+                Keyword("using"),
+                Namespace("System"),
+                Identifier("System"),
+                Operators.Dot,
+                Namespace("Text"),
+                Identifier("Text"),
+                Operators.Dot,
+                Namespace("RegularExpressions"),
+                Identifier("RegularExpressions"),
+                Punctuation.Semicolon,
+                Keyword("class"),
+                Class("C"),
+                Punctuation.OpenCurly,
+                Comment("// class D { }"),
+                Keyword("void"),
+                Method("M"),
+                Punctuation.OpenParen,
+                Punctuation.CloseParen,
+                Punctuation.OpenCurly,
+                Keyword("new"),
+                Class("Regex"),
+                Identifier("Regex"),
+                Punctuation.OpenParen,
+                String("\""),
+                Regex.Grouping("("),
+                Regex.Text("a"),
+                Regex.Grouping(")"),
+                String("\""),
+                Punctuation.CloseParen,
+                Punctuation.Semicolon,
+                Keyword("var"),
+                Keyword("var"),
+                Local("s1"),
+                Operators.Equals,
+                String("\"s1\""),
+                Punctuation.Semicolon,
+                Keyword("var"),
+                Keyword("var"),
+                Local("s2"),
+                Operators.Equals,
+                String("$\""),
+                String("s2"),
+                String("\""),
+                Punctuation.Semicolon,
+                Keyword("var"),
+                Keyword("var"),
+                Local("s3"),
+                Operators.Equals,
+                Verbatim("@\"s3\""),
+                Punctuation.Semicolon,
+                Keyword("var"),
+                Keyword("var"),
+                Local("s4"),
+                Operators.Equals,
+                String(""""
+                    """
+                            s4
+                            """
+                    """"),
+                Punctuation.Semicolon,
+                Punctuation.CloseCurly,
+                Punctuation.CloseCurly,
+            }, actualFormatted);
         }
     }
 }
